Move Pathfinder's methods to Pathfinder.cpp file
